Move membershiplists to bridge/
