+++ lib/bridge/matrix-to-irc.js
@@ -4,7 +4,7 @@ var q = require("q");
 
 var matrixLib = require("../mxlib/matrix");
 var ircLib = require("../irclib/irc");
-var membershiplists = require("../membershiplists");
+var membershiplists = require("./membershiplists");
 var store = require("../store");
 
 var roomModels = require("../models/rooms");
