+++ lib/bridge/membershiplists.js
@@ -3,12 +3,12 @@
 "use strict";
 
 var q = require("q");
-var matrixToIrc = require("./bridge/matrix-to-irc.js");
-var ircLib = require("./irclib/irc.js");
-var matrixLib = require("./mxlib/matrix");
-var MatrixUser = require("./models/users").MatrixUser;
-var store = require("./store");
-var log = require("./logging").get("membershiplists");
+var matrixToIrc = require("./matrix-to-irc.js");
+var ircLib = require("../irclib/irc.js");
+var matrixLib = require("../mxlib/matrix");
+var MatrixUser = require("../models/users").MatrixUser;
+var store = require("../store");
+var log = require("../logging").get("membershiplists");
 
 module.exports.sync = function(server) {
     if (!server.isMembershipListsEnabled()) {
