+++ lib/irc-appservice.js
@@ -9,11 +9,11 @@ var q = require("q");
 
 var matrixToIrc = require("./bridge/matrix-to-irc.js");
 var ircToMatrix = require("./bridge/irc-to-matrix.js");
+var membershiplists = require("./bridge/membershiplists.js");
 var IrcServer = require("./irclib/server.js").IrcServer;
 var ircLib = require("./irclib/irc.js");
 var matrixLib = require("./mxlib/matrix");
 var MatrixUser = require("./models/users").MatrixUser;
-var membershiplists = require("./membershiplists.js");
 var store = require("./store");
 var stats = require("./config/stats");
 var ident = require("./irclib/ident");
