+++ lib/irclib/irc.js
@@ -6,7 +6,7 @@
 var q = require("q");
 var log = require("../logging").get("irc");
 
-var membershiplists = require("../membershiplists.js");
+var membershiplists = require("../bridge/membershiplists.js");
 var pool = require("./client-pool");
 var IrcUser = require("../models/users").IrcUser;
 var MatrixUser = require("../models/users").MatrixUser;
